move egg-related files to own dir
